+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64/FitCloudKit.framework/Headers/FitCloudModels.h
@@ -1,0 +1,15 @@
+//
+//  FitCloudModels.h
+//  FitCloudKit
+//
+//  Created by pcjbird on 1/18/25.
+//  Copyright © 2025 Zero Status. All rights reserved.
+//
+
+#ifndef FitCloudModels_h
+#define FitCloudModels_h
+
+#import <FitCloudKit/FitCloudWorkoutSlot.h>
+
+
+#endif /* FitCloudModels_h */